consolidate classes service testfiles
